+++ client/src/formSchemas/responseErrorSchema.ts
@@ -4,4 +4,5 @@ import { ResponseError } from 'openapi-generator/api/model';
 const ResponseErrorSchema: z.ZodType<ResponseError> = z.object({
   messages: z.string().array().nonempty(),
 });
+
 export default ResponseErrorSchema;
