+++ client/src/pages/sign_up.tsx
@@ -1,18 +1,14 @@
 import type { NextPage } from 'next';
-import { useRouter } from 'next/router';
 import { useForm } from 'react-hook-form';
 import { Box, Text } from '@chakra-ui/react';
 import { zodResolver } from '@hookform/resolvers/zod';
 import InputForm from 'src/components/InputForm';
 import OutlineButton from 'src/components/OutlineButton';
 import { SignUpSchema, SignUpForm } from 'src/formSchemas/signUpSchema';
-import ResponseErrorSchema from 'src/formSchemas/responseErrorSchema';
-import { UserApi, SignUpRequest } from 'openapi-generator/api';
-import axios from 'axios';
-import { setCookie } from 'nookies';
-import useToast from 'src/hooks/toast';
+import useSignUp from 'src/hooks/signUp';
 
 const SignUp: NextPage = () => {
+  const signUp = useSignUp();
   const {
     handleSubmit,
     register,
@@ -20,37 +16,6 @@ const SignUp: NextPage = () => {
   } = useForm<SignUpForm>({
     resolver: zodResolver(SignUpSchema),
   });
-
-  const toast = useToast();
-  const router = useRouter();
-
-  const signUp = async (params: SignUpRequest) => {
-    try {
-      const response = await new UserApi().signUp({
-        name: params.name,
-        email: params.email,
-        password: params.password,
-      });
-      setCookie(null, 'access-token', response.headers['access-token']);
-      setCookie(null, 'uid', response.headers.uid);
-      setCookie(null, 'client', response.headers.client);
-      void router.push('/');
-      toast('success', 'サインアップしました');
-    } catch (error) {
-      if (axios.isAxiosError(error) && error.response !== undefined) {
-        const responseError = ResponseErrorSchema.safeParse(
-          error.response.data
-        );
-        if (responseError.success) {
-          toast(
-            'error',
-            'サインアップに失敗しました',
-            responseError.data.messages.join(`\n`)
-          );
-        }
-      }
-    }
-  };
 
   return (
     <>
